+++ database/migrations/2021_03_15_133242_create_labels_table.php
@@ -17,7 +17,7 @@ class CreateLabelsTable extends Migration
             $table->id();
             $table->string('name');
             $table->text('description')->nullable();
-            $table->bigInteger('creator_id');
+            $table->integer('creator_id');
             $table->foreign('creator_id')->references('id')->on('users');
             $table->timestamps();
         });
